Deleted package.json Angular stuff that wasn't used Reorganized test files to match new src layout Updated test imports (still need to update to ES6 from commonjs)
